Clean-up uplot a bit
